satellite/compensation: make surge percent an int64 for strictcsv
(cherry picked from commit ffe7a3c21195eed99ff3ff2405754132788463d1)

+++ cmd/satellite/main.go
@@ -182,7 +182,7 @@ var (
 		Database     string `help:"satellite database connection string" releaseDefault:"postgres://" devDefault:"postgres://"`
 		Output       string `help:"destination of report output" default:""`
 		Compensation compensation.Config
-		SurgePercent int `help:"surge percent for payments" default:"0"`
+		SurgePercent int64 `help:"surge percent for payments" default:"0"`
 	}
 	recordPeriodCfg struct {
 		Database string `help:"satellite database connection string" releaseDefault:"postgres://" devDefault:"postgres://"`
